1.iOS推送SDK增加音视频录像开关,方便更细粒度的控制录像; 2.iOS推送SDK摄像头demo支持缩放模式下多分辨率选择; 3.iOS推送SDK增加外部h264/aac编码后数据录像接口支持; 4.iOS平台整体优化屏幕采集推送模块，支持推送端横竖屏切换自动适配; 5.iOS播放SDK进一步降低在0缓存情况下的延时; 6.整体优化iOS推送和播放SDK.
1.iOS推送SDK增加音视频录像开关,方便更细粒度的控制录像; 2.iOS推送SDK摄像头demo支持缩放模式下多分辨率选择; 3.iOS推送SDK增加外部h264/aac编码后数据录像接口支持; 4.iOS平台整体优化屏幕采集推送模块，支持推送端横竖屏切换自动适配; 5.iOS播放SDK进一步降低在0缓存情况下的延时; 6.整体优化iOS推送和播放SDK.
