+++ SourceCode/IOS/SmartiOSPublisherV2/SmartiOSPublisher.xcodeproj/xcuserdata/ni.xcuserdatad/xcschemes/SmartiOSPublisher.xcscheme
@@ -45,6 +45,7 @@
       buildConfiguration = "Release"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
+      disableMainThreadChecker = "YES"
       launchStyle = "0"
       useCustomWorkingDirectory = "NO"
       ignoresPersistentStateOnLaunch = "NO"
